requirements.txt
anyio==4.2.0
aws-psycopg2==1.3.8
boto3==1.34.32
botocore==1.34.32
certifi==2023.11.17
charset-normalizer==3.3.2
ecdsa==0.18.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
jmespath==1.0.1
pyasn1==0.5.1
python-dateutil==2.8.2
python-jose==3.3.0
requests==2.31.0
rsa==4.9
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
urllib3==2.0.7
